ref: create confirmation verification code generator
