+++ src/Resources/config/services.xml
@@ -44,6 +44,13 @@
             <argument type="service" id="logger"/>
             <argument type="service" id="emsf.endpoint_manager"/>
         </service>
+
+        <service id="EMS\FormBundle\Contracts\Confirmation\VerificationCodeGeneratorInterface" alias="emsf.service.confirmation.verification_code_generator" />
+        <service id="emsf.service.confirmation.verification_code_generator" class="EMS\FormBundle\Service\Confirmation\VerificationCodeGenerator">
+            <argument type="service" id="EMS\CommonBundle\Contracts\CoreApi\CoreApiInterface"/>
+            <argument type="service" id="session"/>
+        </service>
+
         <service id="EMS\FormBundle\Contracts\EndpointManagerInterface" alias="emsf.endpoint_manager" />
         <service id="emsf.endpoint_manager" class="EMS\FormBundle\Service\Endpoint\EndpointManager">
             <argument>%emsf.endpoints%</argument>
@@ -51,10 +58,9 @@
             <argument type="service" id="logger"/>
         </service>
         <service id="emsf.endpoint.confirmation.http" class="EMS\FormBundle\Service\Confirmation\Endpoint\HttpEndpointType">
-            <argument type="service" id="EMS\CommonBundle\Contracts\CoreApi\CoreApiInterface"/>
             <argument type="service" id="Symfony\Contracts\HttpClient\HttpClientInterface"/>
-            <argument type="service" id="session"/>
             <argument type="service" id="translator"/>
+            <argument type="service" id="EMS\FormBundle\Contracts\Confirmation\VerificationCodeGeneratorInterface"/>
             <tag name="emsf.endpoint.type"/>
         </service>
